Refactor routes into separate modules
