+++ app/routes/routes.py
@@ -52,46 +52,14 @@ ALLOWED_IMPORT_EXTENSIONS = {".txt"}
 
 # If you're not already using a Blueprint for your main routes, create one.
 main = Blueprint('main', __name__)
-location = Blueprint('locations', __name__)
-item = Blueprint('item', __name__)
-transfer = Blueprint('transfer', __name__)
 product = Blueprint('product', __name__)
 customer = Blueprint('customer', __name__)
-invoice = Blueprint('invoice', __name__)
 report = Blueprint('report', __name__)
 purchase = Blueprint('purchase', __name__)
 vendor = Blueprint('vendor', __name__)
 glcode_bp = Blueprint('glcode', __name__)
 
 
-def update_expected_counts(transfer, multiplier=1):
-    """Update expected counts for locations involved in a transfer."""
-    for ti in transfer.transfer_items:
-        from_record = LocationStandItem.query.filter_by(
-            location_id=transfer.from_location_id,
-            item_id=ti.item_id
-        ).first()
-        if not from_record:
-            from_record = LocationStandItem(
-                location_id=transfer.from_location_id,
-                item_id=ti.item_id,
-                expected_count=0
-            )
-            db.session.add(from_record)
-        from_record.expected_count -= multiplier * ti.quantity
-
-        to_record = LocationStandItem.query.filter_by(
-            location_id=transfer.to_location_id,
-            item_id=ti.item_id
-        ).first()
-        if not to_record:
-            to_record = LocationStandItem(
-                location_id=transfer.to_location_id,
-                item_id=ti.item_id,
-                expected_count=0
-            )
-            db.session.add(to_record)
-        to_record.expected_count += multiplier * ti.quantity
 
 
 @main.route('/')
@@ -101,610 +69,10 @@ def home():
     return render_template('transfers/view_transfers.html', user=current_user)
 
 
-@location.route('/locations/add', methods=['GET', 'POST'])
-@login_required
-def add_location():
-    """Create a new location."""
-    form = LocationForm()
-    if form.validate_on_submit():
-        new_location = Location(name=form.name.data)
-        product_ids = [int(pid) for pid in form.products.data.split(',') if pid] if form.products.data else []
-        selected_products = [db.session.get(Product, pid) for pid in product_ids]
-        new_location.products = selected_products
-        db.session.add(new_location)
-        db.session.commit()
 
-        # Add stand sheet items for countable recipe items
-        for product_obj in selected_products:
-            for recipe_item in product_obj.recipe_items:
-                if recipe_item.countable:
-                    exists = LocationStandItem.query.filter_by(
-                        location_id=new_location.id,
-                        item_id=recipe_item.item_id
-                    ).first()
-                    if not exists:
-                        db.session.add(LocationStandItem(location_id=new_location.id, item_id=recipe_item.item_id, expected_count=0))
-        db.session.commit()
-        log_activity(f'Added location {new_location.name}')
-        flash('Location added successfully!')
-        return redirect(url_for('locations.view_locations'))
-    selected_products = []
-    if form.products.data:
-        ids = [int(pid) for pid in form.products.data.split(',') if pid]
-        selected_products = Product.query.filter(Product.id.in_(ids)).all()
-    selected_data = [{'id': p.id, 'name': p.name} for p in selected_products]
-    return render_template('locations/add_location.html', form=form,
-                           selected_products=selected_data)
 
 
-@location.route('/locations/edit/<int:location_id>', methods=['GET', 'POST'])
-@login_required
-def edit_location(location_id):
-    """Edit an existing location."""
-    location = db.session.get(Location, location_id)
-    if location is None:
-        abort(404)
-    form = LocationForm(obj=location)
-    if request.method == 'GET':
-        form.products.data = ','.join(str(p.id) for p in location.products)
 
-    if form.validate_on_submit():
-        location.name = form.name.data
-        product_ids = [int(pid) for pid in form.products.data.split(',') if pid] if form.products.data else []
-        selected_products = [db.session.get(Product, pid) for pid in product_ids]
-        location.products = selected_products
-        db.session.commit()
-
-        # Ensure stand sheet items exist for new products
-        for product_obj in selected_products:
-            for recipe_item in product_obj.recipe_items:
-                if recipe_item.countable:
-                    exists = LocationStandItem.query.filter_by(
-                        location_id=location.id,
-                        item_id=recipe_item.item_id
-                    ).first()
-                    if not exists:
-                        db.session.add(LocationStandItem(location_id=location.id, item_id=recipe_item.item_id, expected_count=0))
-        db.session.commit()
-        log_activity(f'Edited location {location.id}')
-        flash('Location updated successfully.', 'success')
-        return redirect(url_for('locations.edit_location', location_id=location.id))
-
-    # Query for completed transfers to this location
-    transfers_to_location = Transfer.query.filter_by(to_location_id=location_id, completed=True).all()
-
-    selected_data = [{'id': p.id, 'name': p.name} for p in location.products]
-    return render_template('locations/edit_location.html', form=form, location=location,
-                           transfers=transfers_to_location,
-                           selected_products=selected_data)
-
-
-@location.route('/locations/<int:location_id>/stand_sheet')
-@login_required
-def view_stand_sheet(location_id):
-    """Display the expected item counts for a location."""
-    location = db.session.get(Location, location_id)
-    if location is None:
-        abort(404)
-
-    stand_items = []
-    seen = set()
-    for product_obj in location.products:
-        for recipe_item in product_obj.recipe_items:
-            if recipe_item.countable and recipe_item.item_id not in seen:
-                seen.add(recipe_item.item_id)
-                record = LocationStandItem.query.filter_by(
-                    location_id=location_id, item_id=recipe_item.item_id
-                ).first()
-                expected = record.expected_count if record else 0
-                stand_items.append({'item': recipe_item.item, 'expected': expected})
-
-    return render_template('locations/stand_sheet.html', location=location, stand_items=stand_items)
-
-
-@location.route('/locations')
-@login_required
-def view_locations():
-    """List all locations."""
-    locations = Location.query.all()
-    delete_form = DeleteForm()
-    return render_template('locations/view_locations.html', locations=locations, delete_form=delete_form)
-
-
-@location.route('/locations/delete/<int:location_id>', methods=['POST'])
-@login_required
-def delete_location(location_id):
-    """Remove a location from the database."""
-    location = db.session.get(Location, location_id)
-    if location is None:
-        abort(404)
-    db.session.delete(location)
-    db.session.commit()
-    log_activity(f'Deleted location {location.id}')
-    flash('Location deleted successfully!')
-    return redirect(url_for('locations.view_locations'))
-
-
-@item.route('/items')
-@login_required
-def view_items():
-    """Display the inventory item list."""
-    items = Item.query.all()
-    form = ItemForm()
-    return render_template('items/view_items.html', items=items, form=form)
-
-
-@item.route('/items/add', methods=['GET', 'POST'])
-@login_required
-def add_item():
-    """Add a new item to inventory."""
-    form = ItemForm()
-    if form.validate_on_submit():
-        recv_count = sum(1 for uf in form.units if uf.form.name.data and uf.form.receiving_default.data)
-        trans_count = sum(1 for uf in form.units if uf.form.name.data and uf.form.transfer_default.data)
-        if recv_count > 1 or trans_count > 1:
-            flash('Only one unit can be set as receiving and transfer default.', 'error')
-            return render_template('items/add_item.html', form=form)
-        item = Item(
-            name=form.name.data,
-            base_unit=form.base_unit.data,
-            gl_code=form.gl_code.data,
-            gl_code_id=form.gl_code_id.data,
-            purchase_gl_code_id=form.purchase_gl_code.data or None,
-        )
-        db.session.add(item)
-        db.session.commit()
-        receiving_set = False
-        transfer_set = False
-        for uf in form.units:
-            unit_form = uf.form
-            if unit_form.name.data:
-                receiving_default = unit_form.receiving_default.data and not receiving_set
-                transfer_default = unit_form.transfer_default.data and not transfer_set
-                db.session.add(ItemUnit(
-                    item_id=item.id,
-                    name=unit_form.name.data,
-                    factor=float(unit_form.factor.data),
-                    receiving_default=receiving_default,
-                    transfer_default=transfer_default
-                ))
-                if receiving_default:
-                    receiving_set = True
-                if transfer_default:
-                    transfer_set = True
-        db.session.commit()
-        log_activity(f'Added item {item.name}')
-        flash('Item added successfully!')
-        return redirect(url_for('item.view_items'))
-    return render_template('items/add_item.html', form=form)
-
-
-@item.route('/items/edit/<int:item_id>', methods=['GET', 'POST'])
-@login_required
-def edit_item(item_id):
-    """Modify an existing item."""
-    item = db.session.get(Item, item_id)
-    if item is None:
-        abort(404)
-    form = ItemForm(obj=item)
-    if request.method == 'GET':
-        form.gl_code.data = item.gl_code
-        form.gl_code_id.data = item.gl_code_id
-        form.purchase_gl_code.data = item.purchase_gl_code_id
-        for idx, unit in enumerate(item.units):
-            if idx < len(form.units):
-                form.units[idx].form.name.data = unit.name
-                form.units[idx].form.factor.data = unit.factor
-                form.units[idx].form.receiving_default.data = unit.receiving_default
-                form.units[idx].form.transfer_default.data = unit.transfer_default
-            else:
-                form.units.append_entry({
-                    'name': unit.name,
-                    'factor': unit.factor,
-                    'receiving_default': unit.receiving_default,
-                    'transfer_default': unit.transfer_default
-                })
-    if form.validate_on_submit():
-        recv_count = sum(1 for uf in form.units if uf.form.name.data and uf.form.receiving_default.data)
-        trans_count = sum(1 for uf in form.units if uf.form.name.data and uf.form.transfer_default.data)
-        if recv_count > 1 or trans_count > 1:
-            flash('Only one unit can be set as receiving and transfer default.', 'error')
-            return render_template('items/edit_item.html', form=form, item=item)
-        item.name = form.name.data
-        item.base_unit = form.base_unit.data
-        item.gl_code = form.gl_code.data
-        item.gl_code_id = form.gl_code_id.data
-        item.purchase_gl_code_id = form.purchase_gl_code.data or None
-        ItemUnit.query.filter_by(item_id=item.id).delete()
-        receiving_set = False
-        transfer_set = False
-        for uf in form.units:
-            unit_form = uf.form
-            if unit_form.name.data:
-                receiving_default = unit_form.receiving_default.data and not receiving_set
-                transfer_default = unit_form.transfer_default.data and not transfer_set
-                db.session.add(ItemUnit(
-                    item_id=item.id,
-                    name=unit_form.name.data,
-                    factor=float(unit_form.factor.data),
-                    receiving_default=receiving_default,
-                    transfer_default=transfer_default
-                ))
-                if receiving_default:
-                    receiving_set = True
-                if transfer_default:
-                    transfer_set = True
-        db.session.commit()
-        log_activity(f'Edited item {item.id}')
-        flash('Item updated successfully!')
-        return redirect(url_for('item.view_items'))
-    return render_template('items/edit_item.html', form=form, item=item)
-
-
-@item.route('/items/delete/<int:item_id>', methods=['POST'])
-@login_required
-def delete_item(item_id):
-    """Delete an item from the catalog."""
-    item = db.session.get(Item, item_id)
-    if item is None:
-        abort(404)
-    db.session.delete(item)
-    db.session.commit()
-    log_activity(f'Deleted item {item.id}')
-    flash('Item deleted successfully!')
-    return redirect(url_for('item.view_items'))
-
-
-@item.route('/items/bulk_delete', methods=['POST'])
-@login_required
-def bulk_delete_items():
-    """Delete multiple items in one request."""
-    item_ids = request.form.getlist('item_ids')
-    if item_ids:
-        Item.query.filter(Item.id.in_(item_ids)).delete(synchronize_session='fetch')
-        db.session.commit()
-        log_activity(f'Bulk deleted items {",".join(item_ids)}')
-        flash('Selected items have been deleted.', 'success')
-    else:
-        flash('No items selected.', 'warning')
-    return redirect(url_for('item.view_items'))
-
-
-@transfer.route('/transfers', methods=['GET'])
-@login_required
-def view_transfers():
-    """Show transfers with optional filtering."""
-    filter_option = request.args.get('filter', 'not_completed')
-    transfer_id = request.args.get('transfer_id', '', type=int)  # Optional: Search by Transfer ID
-    from_location_name = request.args.get('from_location', '')  # Optional: Search by From Location
-    to_location_name = request.args.get('to_location', '')  # Optional: Search by To Location
-
-    query = Transfer.query
-    if transfer_id != '':
-        query = query.filter(Transfer.id == transfer_id)
-
-    if from_location_name != '':
-        query = query.join(Location, Transfer.from_location_id == Location.id).filter(
-            Location.name.ilike(f"%{from_location_name}%"))
-
-    if to_location_name != '':
-        query = query.join(Location, Transfer.to_location_id == Location.id).filter(
-            Location.name.ilike(f"%{to_location_name}%"))
-
-    if filter_option == 'completed':
-        transfers = query.filter(Transfer.completed == True).all()
-    elif filter_option == 'not_completed':
-        transfers = query.filter(Transfer.completed == False).all()
-    else:
-        transfers = query.all()
-
-    form = TransferForm()  # Assuming you're using it for something like a filter form on the page
-    return render_template('transfers/view_transfers.html', transfers=transfers, form=form)
-
-
-@transfer.route('/transfers/add', methods=['GET', 'POST'])
-@login_required
-def add_transfer():
-    """Create a transfer between locations."""
-    form = TransferForm()
-    if form.validate_on_submit():
-        transfer = Transfer(
-            from_location_id=form.from_location_id.data,
-            to_location_id=form.to_location_id.data,
-            user_id=current_user.id
-        )
-        db.session.add(transfer)
-        # Dynamically determine the number of items added
-        items = [key for key in request.form.keys() if key.startswith('items-') and key.endswith('-item')]
-        for item_field in items:
-            index = item_field.split('-')[1]
-            item_id = request.form.get(f'items-{index}-item')
-            quantity = request.form.get(f'items-{index}-quantity', type=float)
-            unit_id = request.form.get(f'items-{index}-unit', type=int)
-            if item_id:
-                item = db.session.get(Item, item_id)
-                if item and quantity is not None:
-                    factor = 1
-                    if unit_id:
-                        unit = db.session.get(ItemUnit, unit_id)
-                        if unit:
-                            factor = unit.factor
-                    transfer_item = TransferItem(
-                        transfer_id=transfer.id,
-                        item_id=item.id,
-                        quantity=quantity * factor
-                    )
-                    db.session.add(transfer_item)
-        db.session.commit()
-        log_activity(f'Added transfer {transfer.id}')
-
-        socketio.emit('new_transfer', {'message': 'New transfer added'})
-
-        flash('Transfer added successfully!', 'success')
-        return redirect(url_for('transfer.view_transfers'))
-    elif form.errors:
-        flash('There was an error submitting the transfer.', 'error')
-
-    return render_template('transfers/add_transfer.html', form=form)
-
-
-@transfer.route('/transfers/edit/<int:transfer_id>', methods=['GET', 'POST'])
-@login_required
-def edit_transfer(transfer_id):
-    """Update an existing transfer."""
-    transfer = db.session.get(Transfer, transfer_id)
-    if transfer is None:
-        abort(404)
-    form = TransferForm(obj=transfer)
-
-    if form.validate_on_submit():
-        transfer.from_location_id = form.from_location_id.data
-        transfer.to_location_id = form.to_location_id.data
-
-        # Clear existing TransferItem entries
-        TransferItem.query.filter_by(transfer_id=transfer.id).delete()
-
-        # Dynamically determine the number of items added, similar to the "add" route
-        items = [key for key in request.form.keys() if key.startswith('items-') and key.endswith('-item')]
-        for item_field in items:
-            index = item_field.split('-')[1]
-            item_id = request.form.get(f'items-{index}-item')
-            quantity = request.form.get(f'items-{index}-quantity', type=float)
-            unit_id = request.form.get(f'items-{index}-unit', type=int)
-            if item_id and quantity is not None:  # Ensure both are provided and valid
-                factor = 1
-                if unit_id:
-                    unit = db.session.get(ItemUnit, unit_id)
-                    if unit:
-                        factor = unit.factor
-                new_transfer_item = TransferItem(
-                    transfer_id=transfer.id,
-                    item_id=int(item_id),
-                    quantity=quantity * factor
-                )
-                db.session.add(new_transfer_item)
-
-        db.session.commit()
-        log_activity(f'Edited transfer {transfer.id}')
-        flash('Transfer updated successfully!', 'success')
-        return redirect(url_for('transfer.view_transfers'))
-    elif form.errors:
-        flash('There was an error submitting the transfer.', 'error')
-
-    # For GET requests or if the form doesn't validate, pass existing items to the template
-    items = [{"id": item.item_id, "name": item.item.name, "quantity": item.quantity} for item in
-             transfer.transfer_items]
-    return render_template('transfers/edit_transfer.html', form=form, transfer=transfer, items=items)
-
-
-@transfer.route('/transfers/delete/<int:transfer_id>', methods=['POST'])
-@login_required
-def delete_transfer(transfer_id):
-    """Permanently remove a transfer."""
-    transfer = db.session.get(Transfer, transfer_id)
-    if transfer is None:
-        abort(404)
-    if transfer.completed:
-        update_expected_counts(transfer, multiplier=-1)
-    db.session.delete(transfer)
-    db.session.commit()
-    log_activity(f'Deleted transfer {transfer.id}')
-    flash('Transfer deleted successfully!', 'success')
-    return redirect(url_for('transfer.view_transfers'))
-
-
-@transfer.route('/transfers/complete/<int:transfer_id>', methods=['GET'])
-@login_required
-def complete_transfer(transfer_id):
-    """Mark a transfer as completed."""
-    transfer = db.session.get(Transfer, transfer_id)
-    if transfer is None:
-        abort(404)
-    transfer.completed = True
-    update_expected_counts(transfer, multiplier=1)
-    db.session.commit()
-    log_activity(f'Completed transfer {transfer.id}')
-    flash('Transfer marked as complete!', 'success')
-    return redirect(url_for('transfer.view_transfers'))
-
-
-@transfer.route('/transfers/uncomplete/<int:transfer_id>', methods=['GET'])
-@login_required
-def uncomplete_transfer(transfer_id):
-    """Revert a transfer to not completed."""
-    transfer = db.session.get(Transfer, transfer_id)
-    if transfer is None:
-        abort(404)
-    transfer.completed = False
-    update_expected_counts(transfer, multiplier=-1)
-    db.session.commit()
-    log_activity(f'Uncompleted transfer {transfer.id}')
-    flash('Transfer marked as not completed.', 'success')
-    return redirect(url_for('transfer.view_transfers'))
-
-
-@transfer.route('/transfers/view/<int:transfer_id>', methods=['GET'])
-@login_required
-def view_transfer(transfer_id):
-    """Show details for a single transfer."""
-    transfer = db.session.get(Transfer, transfer_id)
-    if transfer is None:
-        abort(404)
-    transfer_items = TransferItem.query.filter_by(transfer_id=transfer.id).all()
-    return render_template('transfers/view_transfer.html', transfer=transfer, transfer_items=transfer_items)
-
-
-@item.route('/items/search', methods=['GET'])
-@login_required
-def search_items():
-    """Search items by name for autocomplete fields."""
-    search_term = request.args.get('term', '')
-    items = Item.query.filter(Item.name.ilike(f'%{search_term}%')).all()
-    items_data = [{'id': item.id, 'name': item.name} for item in items]  # Create a list of dicts
-    return jsonify(items_data)
-
-
-@item.route('/items/<int:item_id>/units')
-@login_required
-def item_units(item_id):
-    """Return unit options for an item."""
-    item = db.session.get(Item, item_id)
-    if item is None:
-        abort(404)
-    data = [
-        {
-            'id': u.id,
-            'name': u.name,
-            'factor': u.factor,
-            'receiving_default': u.receiving_default,
-            'transfer_default': u.transfer_default,
-        }
-        for u in item.units
-    ]
-    return jsonify(data)
-
-
-@item.route('/items/<int:item_id>/last_cost')
-@login_required
-def item_last_cost(item_id):
-    """Return the last recorded cost for an item."""
-    unit_id = request.args.get('unit_id', type=int)
-    item = db.session.get(Item, item_id)
-    if item is None:
-        abort(404)
-    factor = 1.0
-    if unit_id:
-        unit = db.session.get(ItemUnit, unit_id)
-        if unit:
-            factor = unit.factor
-    return jsonify({'cost': (item.cost or 0.0) * factor})
-
-
-@item.route('/import_items', methods=['GET', 'POST'])
-@login_required
-def import_items():
-    """Bulk import items from a text file."""
-    form = ImportItemsForm()
-    if form.validate_on_submit():
-        from run import app
-
-        file = form.file.data
-        filename = secure_filename(file.filename)
-        ext = os.path.splitext(filename)[1].lower()
-        file.seek(0, os.SEEK_END)
-        size = file.tell()
-        file.seek(0)
-        if ext not in ALLOWED_IMPORT_EXTENSIONS:
-            flash('Only .txt files are allowed.', 'error')
-            return redirect(url_for('item.import_items'))
-        if size > MAX_IMPORT_SIZE:
-            flash('File is too large.', 'error')
-            return redirect(url_for('item.import_items'))
-        filepath = os.path.join(app.config['UPLOAD_FOLDER'], filename)
-        file.save(filepath)
-
-        # Parse the file
-        with open(filepath, 'r') as file:
-            for line in file:
-                item_name = line.strip()
-                if item_name:
-                    # Check if item already exists to avoid duplicates
-                    existing_item = Item.query.filter_by(name=item_name).first()
-                    if not existing_item:
-                        # Create a new item instance and add it to the database
-                        new_item = Item(name=item_name)
-                        db.session.add(new_item)
-        db.session.commit()
-        log_activity('Imported items from file')
-
-        flash('Items imported successfully.', 'success')
-        return redirect(url_for('item.import_items'))
-
-    return render_template('items/import_items.html', form=form)
-
-
-@transfer.route('/transfers/generate_report', methods=['GET', 'POST'])
-def generate_report():
-    """Generate a transfer summary over a date range."""
-    form = DateRangeForm()
-    if form.validate_on_submit():
-        start_datetime = form.start_datetime.data
-        end_datetime = form.end_datetime.data
-
-        # Alias for "from" and "to" locations
-        from_location = db.aliased(Location)
-        to_location = db.aliased(Location)
-
-        aggregated_transfers = db.session.query(
-            from_location.name.label('from_location_name'),
-            to_location.name.label('to_location_name'),
-            Item.name.label('item_name'),
-            func.sum(TransferItem.quantity).label('total_quantity')
-        ).select_from(Transfer) \
-            .join(TransferItem, Transfer.id == TransferItem.transfer_id) \
-            .join(Item, TransferItem.item_id == Item.id) \
-            .join(from_location, Transfer.from_location_id == from_location.id) \
-            .join(to_location, Transfer.to_location_id == to_location.id) \
-            .filter(
-            Transfer.completed == True,
-            Transfer.date_created >= start_datetime,
-            Transfer.date_created <= end_datetime
-        ) \
-            .group_by(
-            from_location.name,
-            to_location.name,
-            Item.name
-        ) \
-            .order_by(
-            from_location.name,
-            to_location.name,
-            Item.name
-        ) \
-            .all()
-
-        # Process the results for display or session storage
-        session['aggregated_transfers'] = [{
-            'from_location_name': result[0],
-            'to_location_name': result[1],
-            'item_name': result[2],
-            'total_quantity': result[3]
-        } for result in aggregated_transfers]
-
-        # Store start and end date/time in session for use in the report
-        session['report_start_datetime'] = start_datetime.strftime('%Y-%m-%d %H:%M')
-        session['report_end_datetime'] = end_datetime.strftime('%Y-%m-%d %H:%M')
-
-        flash('Transfer report generated successfully.', 'success')
-        return redirect(url_for('transfer.view_report'))
-
-    return render_template('transfers/generate_report.html', form=form)
-
-
-@transfer.route('/transfers/report')
-def view_report():
-    """Display the previously generated transfer report."""
-    aggregated_transfers = session.get('aggregated_transfers', [])
-    return render_template('transfers/view_report.html', aggregated_transfers=aggregated_transfers)
 
 
 @product.route('/products')
@@ -1094,191 +462,6 @@ def search_products():
     return jsonify(product_data)
 
 
-@invoice.route('/create_invoice', methods=['GET', 'POST'])
-@login_required
-def create_invoice():
-    """Create a sales invoice."""
-    form = InvoiceForm()
-    form.customer.choices = [(c.id, f"{c.first_name} {c.last_name}") for c in Customer.query.all()]
-
-    if form.validate_on_submit():
-        customer = db.session.get(Customer, form.customer.data)
-        if customer is None:
-            abort(404)
-        today = datetime.now().strftime('%d%m%y')
-        count = Invoice.query.filter(
-            func.date(Invoice.date_created) == func.current_date(),
-            Invoice.customer_id == customer.id
-        ).count() + 1
-        invoice_id = f"{customer.first_name[0]}{customer.last_name[0]}{customer.id}{today}{count:02}"
-
-        invoice = Invoice(
-            id=invoice_id,
-            customer_id=customer.id,
-            user_id=current_user.id
-        )
-        db.session.add(invoice)
-
-        product_data = form.products.data.removesuffix(":").split(':')
-
-        for entry in product_data:
-            try:
-                product_name, quantity, override_gst, override_pst = entry.split('?')
-                product = Product.query.filter_by(name=product_name).first()
-
-                if product:
-                    quantity = float(quantity)
-                    unit_price = product.price
-                    line_subtotal = quantity * unit_price
-
-                    # Parse overrides correctly (can be 0, 1, or empty string)
-                    override_gst = None if override_gst == '' else bool(int(override_gst))
-                    override_pst = None if override_pst == '' else bool(int(override_pst))
-
-                    # Apply tax rules
-                    apply_gst = override_gst if override_gst is not None else not customer.gst_exempt
-                    apply_pst = override_pst if override_pst is not None else not customer.pst_exempt
-
-                    line_gst = line_subtotal * 0.05 if apply_gst else 0
-                    line_pst = line_subtotal * 0.07 if apply_pst else 0
-
-                    invoice_product = InvoiceProduct(
-                        invoice_id=invoice.id,
-                        product_id=product.id,
-                        quantity=quantity,
-                        override_gst=override_gst,
-                        override_pst=override_pst,
-                        unit_price=unit_price,
-                        line_subtotal=line_subtotal,
-                        line_gst=line_gst,
-                        line_pst=line_pst
-                    )
-                    db.session.add(invoice_product)
-
-                    # Reduce product inventory
-                    product.quantity = (product.quantity or 0) - quantity
-
-                    # Reduce item inventories based on recipe
-                    for recipe_item in product.recipe_items:
-                        item = recipe_item.item
-                        item.quantity = (item.quantity or 0) - (recipe_item.quantity * quantity)
-                else:
-                    flash(f"Product '{product_name}' not found.", 'danger')
-
-            except ValueError:
-                flash(f"Invalid product data format: '{entry}'", 'danger')
-
-        db.session.commit()
-        log_activity(f'Created invoice {invoice.id}')
-        flash('Invoice created successfully!', 'success')
-        return redirect(url_for('invoice.view_invoices'))
-
-    return render_template('create_invoice.html', form=form)
-
-
-@invoice.route('/delete_invoice/<invoice_id>', methods=['GET'])
-@login_required
-def delete_invoice(invoice_id):
-    """Delete an invoice and its lines."""
-    # Retrieve the invoice object from the database
-    invoice = db.session.get(Invoice, invoice_id)
-    if invoice is None:
-        abort(404)
-    # Delete the invoice from the database
-    db.session.delete(invoice)
-    db.session.commit()
-    log_activity(f'Deleted invoice {invoice.id}')
-    flash('Invoice deleted successfully!', 'success')
-    # Redirect the user to the home page or any other appropriate page
-    return redirect(url_for('invoice.view_invoices'))
-
-
-@invoice.route('/view_invoice/<invoice_id>', methods=['GET'])
-@login_required
-def view_invoice(invoice_id):
-    """Render an invoice for viewing."""
-    invoice = db.session.get(Invoice, invoice_id)
-    if invoice is None:
-        abort(404)
-
-    subtotal = 0
-    gst_total = 0
-    pst_total = 0
-
-    for invoice_product in invoice.products:
-        # Use stored values instead of recalculating from current product price
-        line_total = invoice_product.line_subtotal
-        subtotal += line_total
-        gst_total += invoice_product.line_gst
-        pst_total += invoice_product.line_pst
-
-    total = subtotal + gst_total + pst_total
-
-    return render_template(
-        'view_invoice.html',
-        invoice=invoice,
-        subtotal=subtotal,
-        gst=gst_total,
-        pst=pst_total,
-        total=total,
-        GST=GST
-    )
-
-@invoice.route('/get_customer_tax_status/<int:customer_id>')
-@login_required
-def get_customer_tax_status(customer_id):
-    """Return GST and PST exemptions for a customer."""
-    customer = db.session.get(Customer, customer_id)
-    if customer is None:
-        abort(404)
-    return {
-        "gst_exempt": customer.gst_exempt,
-        "pst_exempt": customer.pst_exempt
-    }
-
-
-@invoice.route('/view_invoices', methods=['GET', 'POST'])
-@login_required
-def view_invoices():
-    """List invoices with optional filters."""
-    form = InvoiceFilterForm()
-    form.vendor_id.choices = [(-1, 'All')] + [
-        (c.id, f"{c.first_name} {c.last_name}") for c in Customer.query.all()
-    ]
-
-    # Determine filter values from form submission or query params
-    if form.validate_on_submit():
-        invoice_id = form.invoice_id.data
-        vendor_id = form.vendor_id.data
-        start_date = form.start_date.data
-        end_date = form.end_date.data
-    else:
-        invoice_id = request.args.get('invoice_id', '')
-        vendor_id = request.args.get('vendor_id', type=int)
-        start_date_str = request.args.get('start_date')
-        end_date_str = request.args.get('end_date')
-        start_date = datetime.fromisoformat(start_date_str) if start_date_str else None
-        end_date = datetime.fromisoformat(end_date_str) if end_date_str else None
-        form.invoice_id.data = invoice_id
-        if vendor_id is not None:
-            form.vendor_id.data = vendor_id
-        if start_date:
-            form.start_date.data = start_date
-        if end_date:
-            form.end_date.data = end_date
-
-    query = Invoice.query
-    if invoice_id:
-        query = query.filter(Invoice.id.ilike(f"%{invoice_id}%"))
-    if vendor_id and vendor_id != -1:
-        query = query.filter(Invoice.customer_id == vendor_id)
-    if start_date:
-        query = query.filter(Invoice.date_created >= datetime.combine(start_date, datetime.min.time()))
-    if end_date:
-        query = query.filter(Invoice.date_created <= datetime.combine(end_date, datetime.max.time()))
-
-    invoices = query.order_by(Invoice.date_created.desc()).all()
-    return render_template('view_invoices.html', invoices=invoices, form=form)
 
 @report.route('/reports/vendor-invoices', methods=['GET', 'POST'])
 def vendor_invoice_report():
